requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-import-export==2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pkg-resources==0.0.0
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
